Add cesm catalog
