+++ intake_cesm/__init__.py
@@ -5,4 +5,4 @@ from ._version import get_versions
 __version__ = get_versions()["version"]
 del get_versions
 
-from .cat import CesmCatalog
+from .cesm_cat import CesmCatalog
