+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.black]
+line-length = 100
+include = '\.pyi?$'
+exclude = '''
+/(
+    \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | _build
+  | buck-out
+  | build
+  | dist
+  | docs
+  | versioneer.py
+  | intake_cesm/_version.py
+)/
+'''
