+++ setup.cfg
@@ -4,7 +4,7 @@ universal = 1
 [flake8]
 exclude = docs, versioneer.py, intake_cesm/_version.py
 ignore = E203,E266,E501,W503,F401,E722,E402,C901
-max-line-length = 88
+max-line-length = 100
 max-complexity = 18
 select = B,C,E,F,W,T4,B9
